Move classes; Extract methods; Extract classes
