+++ app/src/main/java/com/edu/uni/augsburg/uniatron/ui/MainActivity.java
@@ -19,7 +19,6 @@ import android.view.MenuItem;
 
 import com.afollestad.materialdialogs.MaterialDialog;
 import com.edu.uni.augsburg.uniatron.R;
-import com.edu.uni.augsburg.uniatron.domain.util.DateFormatter;
 import com.edu.uni.augsburg.uniatron.notification.NotificationChannels;
 import com.edu.uni.augsburg.uniatron.service.StickyAppService;
 import com.edu.uni.augsburg.uniatron.ui.about.AboutActivity;
@@ -30,6 +29,7 @@ import com.edu.uni.augsburg.uniatron.ui.card.timeaccount.TimeAccountViewModel;
 import com.edu.uni.augsburg.uniatron.ui.onboarding.OnBoardingActivity;
 import com.edu.uni.augsburg.uniatron.ui.setting.SettingActivity;
 import com.edu.uni.augsburg.uniatron.ui.shop.TimeCreditShopActivity;
+import com.edu.uni.augsburg.uniatron.ui.util.DateFormatter;
 
 import java.util.Calendar;
 import java.util.Date;
