+++ app/src/main/java/com/edu/uni/augsburg/uniatron/ui/card/CardPriority.java
@@ -1,0 +1,29 @@
+package com.edu.uni.augsburg.uniatron.ui.card;
+
+/**
+ * The priority to display the card.
+ *
+ * @author Fabio Hellmann
+ */
+public enum CardPriority {
+    /**
+     * Very high priority.
+     */
+    VERY_HIGH,
+    /**
+     * High priority.
+     */
+    HIGH,
+    /**
+     * Normal priority.
+     */
+    NORMAL,
+    /**
+     * Low priority.
+     */
+    LOW,
+    /**
+     * Very low priority.
+     */
+    VERY_LOW
+}
