+++ app/src/main/java/com/edu/uni/augsburg/uniatron/ui/card/EmptyCard.java
@@ -8,7 +8,6 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import com.edu.uni.augsburg.uniatron.R;
-import com.edu.uni.augsburg.uniatron.ui.CardViewHolder;
 
 /**
  * An empty card.
