+++ app/src/main/java/com/edu/uni/augsburg/uniatron/ui/card/appusage/AppUsageCard.java
@@ -1,26 +1,23 @@
 package com.edu.uni.augsburg.uniatron.ui.card.appusage;
 
 import android.content.Context;
-import android.graphics.drawable.Drawable;
 import android.support.annotation.NonNull;
 import android.support.design.button.MaterialButton;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.StaggeredGridLayoutManager;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ImageView;
 import android.widget.TextView;
 
 import com.annimon.stream.Stream;
 import com.edu.uni.augsburg.uniatron.R;
-import com.edu.uni.augsburg.uniatron.ui.CardViewHolder;
+import com.edu.uni.augsburg.uniatron.ui.card.CardHelper;
+import com.edu.uni.augsburg.uniatron.ui.card.CardViewHolder;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.Locale;
 import java.util.concurrent.TimeUnit;
 
 import butterknife.BindView;
@@ -34,7 +31,6 @@ import butterknife.OnClick;
  */
 public class AppUsageCard implements CardViewHolder {
 
-    private static final int TYPE = 10;
     private final List<AppUsageViewItem> mAppUsageList = new ArrayList<>();
 
     /**
@@ -57,10 +53,7 @@ public class AppUsageCard implements CardViewHolder {
     @Override
     public void onBindView(@NonNull final Context context,
                            @NonNull final RecyclerView.ViewHolder viewHolder) {
-        final StaggeredGridLayoutManager.LayoutParams layoutParams =
-                (StaggeredGridLayoutManager.LayoutParams) viewHolder.itemView.getLayoutParams();
-        layoutParams.setFullSpan(true);
-
+        CardHelper.setFullSpan(viewHolder);
         final ViewHolder holder = (ViewHolder) viewHolder;
 
         final LinearLayoutManager layout = new LinearLayoutManager(context);
@@ -71,11 +64,6 @@ public class AppUsageCard implements CardViewHolder {
         final long totalAppUsage = TimeUnit.SECONDS.convert(getTotalAppUsage(), TimeUnit.MILLISECONDS);
         holder.mTextAppUsageTotal.setText(context.getString(R.string.time_in_minutes, totalAppUsage / 60, totalAppUsage % 60));
         holder.mButtonShowAll.setText(R.string.show_all);
-    }
-
-    @Override
-    public int getType() {
-        return TYPE;
     }
 
     @Override
@@ -108,85 +96,14 @@ public class AppUsageCard implements CardViewHolder {
         public void onShowAllClicked() {
             final AppUsageAdapter adapter = (AppUsageAdapter) mRecyclerView.getAdapter();
             if (adapter != null) {
-                if (adapter.mShowSmallVisibleAmount) {
+                if (adapter.isShowSmallVisibleAmount()) {
                     mButtonShowAll.setText(R.string.show_only_top_five);
                 } else {
                     mButtonShowAll.setText(R.string.show_all);
                 }
-                adapter.setShowSomeItems(!adapter.mShowSmallVisibleAmount);
+                adapter.setShowSomeItems(!adapter.isShowSmallVisibleAmount());
                 adapter.notifyDataSetChanged();
             }
-        }
-    }
-
-    static final class AppUsageAdapter extends RecyclerView.Adapter<AppUsageCard.ViewHolderListItem> {
-        private static final int SMALL_VISIBLE_AMOUNT = 5;
-        private final Context mContext;
-        private final List<AppUsageViewItem> mAppData;
-        private boolean mShowSmallVisibleAmount = true;
-
-        AppUsageAdapter(@NonNull final Context context,
-                        @NonNull final List<AppUsageViewItem> appData) {
-            super();
-            mContext = context;
-            mAppData = appData;
-        }
-
-        void setShowSomeItems(final boolean visible) {
-            mShowSmallVisibleAmount = visible;
-        }
-
-        @NonNull
-        @Override
-        public ViewHolderListItem onCreateViewHolder(@NonNull final ViewGroup viewGroup,
-                                                     final int position) {
-            return new ViewHolderListItem(LayoutInflater.from(mContext)
-                    .inflate(R.layout.card_app_usage_item, viewGroup, false));
-        }
-
-        @Override
-        public void onBindViewHolder(@NonNull final ViewHolderListItem viewHolderListItem,
-                                     final int position) {
-            final AppUsageViewItem item = mAppData.get(position);
-
-            final String appLabel = item.getAppLabel();
-            final Drawable appIcon = item.getAppIcon();
-            final long usageTime = TimeUnit.SECONDS.convert(item.getApplicationUsage(), TimeUnit.MILLISECONDS);
-            final double usageTimePercent = item.getApplicationUsagePercent();
-
-            viewHolderListItem.mTextAppName.setText(appLabel);
-            viewHolderListItem.mImageAppIcon.setImageDrawable(appIcon);
-            viewHolderListItem.mTextAppUsage.setText(mContext.getString(
-                    R.string.app_usage_time,
-                    usageTime / 60,
-                    usageTime % 60
-            ));
-            viewHolderListItem.mTextAppUsagePercent.setText(String.format(
-                    Locale.getDefault(),
-                    "%d %%",
-                    Math.round(usageTimePercent)
-            ));
-        }
-
-        @Override
-        public int getItemCount() {
-            return mAppData.size() > SMALL_VISIBLE_AMOUNT && mShowSmallVisibleAmount ? SMALL_VISIBLE_AMOUNT : mAppData.size();
-        }
-    }
-
-    static final class ViewHolderListItem extends RecyclerView.ViewHolder {
-        @BindView(R.id.imageAppIcon)
-        ImageView mImageAppIcon;
-        @BindView(R.id.textAppName)
-        TextView mTextAppName;
-        @BindView(R.id.textAppUsage)
-        TextView mTextAppUsage;
-        @BindView(R.id.textAppUsagePercent)
-        TextView mTextAppUsagePercent;
-
-        ViewHolderListItem(@NonNull final View itemView) {
-            super(itemView);
-            ButterKnife.bind(this, itemView);
         }
     }
 }
