+++ app/src/main/java/com/edu/uni/augsburg/uniatron/ui/card/coinbag/CoinBagViewModel.java
@@ -10,7 +10,7 @@ import android.support.annotation.NonNull;
 import com.edu.uni.augsburg.uniatron.MainApplication;
 import com.edu.uni.augsburg.uniatron.domain.DataRepository;
 import com.edu.uni.augsburg.uniatron.domain.util.DateConverter;
-import com.edu.uni.augsburg.uniatron.ui.CardViewModel;
+import com.edu.uni.augsburg.uniatron.ui.card.CardViewModel;
 import com.edu.uni.augsburg.uniatron.ui.card.DateCache;
 
 import java.util.Date;
