+++ app/src/main/java/com/edu/uni/augsburg/uniatron/ui/card/timeaccount/TimeAccountViewModel.java
@@ -11,7 +11,7 @@ import com.edu.uni.augsburg.uniatron.MainApplication;
 import com.edu.uni.augsburg.uniatron.SharedPreferencesHandler;
 import com.edu.uni.augsburg.uniatron.domain.DataRepository;
 import com.edu.uni.augsburg.uniatron.domain.util.DateConverter;
-import com.edu.uni.augsburg.uniatron.ui.CardViewModel;
+import com.edu.uni.augsburg.uniatron.ui.card.CardViewModel;
 import com.edu.uni.augsburg.uniatron.ui.card.DateCache;
 
 import java.util.Date;
