+++ app/src/test/java/com/edu/uni/augsburg/uniatron/ui/util/DateFormatterTest.java
@@ -1,4 +1,4 @@
-package com.edu.uni.augsburg.uniatron.domain.util;
+package com.edu.uni.augsburg.uniatron.ui.util;
 
 import org.junit.Test;
